analyzer/requirements.txt
fastapi==0.110.0
uvicorn==0.25.0
psycopg2-binary==2.9.9
pandas==2.2.0
python-dotenv==1.0.1